Move 'standart' preference pages (Appearance, Plugins) to the same package where 'PreferencesPresenter' is
