+++ ide/che-core-ide-app/src/main/java/org/eclipse/che/ide/core/inject/CoreGinModule.java
@@ -103,9 +103,9 @@ import org.eclipse.che.ide.core.Component;
 import org.eclipse.che.ide.core.StandardComponentInitializer;
 import org.eclipse.che.ide.core.editor.EditorAgentImpl;
 import org.eclipse.che.ide.core.editor.EditorRegistryImpl;
-import org.eclipse.che.ide.extension.ExtensionManagerPresenter;
-import org.eclipse.che.ide.extension.ExtensionManagerView;
-import org.eclipse.che.ide.extension.ExtensionManagerViewImpl;
+import org.eclipse.che.ide.preferences.pages.extensions.ExtensionManagerPresenter;
+import org.eclipse.che.ide.preferences.pages.extensions.ExtensionManagerView;
+import org.eclipse.che.ide.preferences.pages.extensions.ExtensionManagerViewImpl;
 import org.eclipse.che.ide.filetypes.FileTypeRegistryImpl;
 import org.eclipse.che.ide.hotkeys.dialog.HotKeysDialogView;
 import org.eclipse.che.ide.hotkeys.dialog.HotKeysDialogViewImpl;
@@ -162,9 +162,9 @@ import org.eclipse.che.ide.settings.common.SettingsPagePresenter;
 import org.eclipse.che.ide.statepersistance.OpenedFilesPersistenceComponent;
 import org.eclipse.che.ide.statepersistance.PersistenceComponent;
 import org.eclipse.che.ide.statepersistance.ShowHiddenFilesPersistenceComponent;
-import org.eclipse.che.ide.theme.AppearancePresenter;
-import org.eclipse.che.ide.theme.AppearanceView;
-import org.eclipse.che.ide.theme.AppearanceViewImpl;
+import org.eclipse.che.ide.preferences.pages.appearance.AppearancePresenter;
+import org.eclipse.che.ide.preferences.pages.appearance.AppearanceView;
+import org.eclipse.che.ide.preferences.pages.appearance.AppearanceViewImpl;
 import org.eclipse.che.ide.theme.DarkTheme;
 import org.eclipse.che.ide.theme.LightTheme;
 import org.eclipse.che.ide.theme.ThemeAgentImpl;
