+++ ide/che-core-ide-app/src/main/java/org/eclipse/che/ide/preferences/pages/appearance/AppearancePresenter.java
@@ -8,7 +8,7 @@
  * Contributors:
  *   Codenvy, S.A. - initial API and implementation
  *******************************************************************************/
-package org.eclipse.che.ide.theme;
+package org.eclipse.che.ide.preferences.pages.appearance;
 
 import org.eclipse.che.ide.CoreLocalizationConstant;
