+++ ide/che-core-ide-app/src/main/java/org/eclipse/che/ide/preferences/pages/appearance/AppearanceViewImpl.ui.xml
