+++ ide/che-core-ide-app/src/main/java/org/eclipse/che/ide/preferences/pages/extensions/ExtensionManagerView.java
@@ -8,22 +8,18 @@
  * Contributors:
  *   Codenvy, S.A. - initial API and implementation
  *******************************************************************************/
-package org.eclipse.che.ide.theme;
+package org.eclipse.che.ide.preferences.pages.extensions;
 
+import org.eclipse.che.ide.api.extension.ExtensionDescription;
 import org.eclipse.che.ide.api.mvp.View;
-import org.eclipse.che.ide.api.theme.Theme;
 
 import java.util.List;
 
-/**
- * @author Evgen Vidolob
- */
-public interface AppearanceView extends View<AppearanceView.ActionDelegate> {
+/** @author Evgen Vidolob */
+public interface ExtensionManagerView extends View<ExtensionManagerView.ActionDelegate> {
 
-    void setThemes(List<Theme> themes, String currentThemeId);
+    void setExtensions(List<ExtensionDescription> extensions);
 
-    public interface ActionDelegate {
-
-        void themeSelected(String themeId);
+    interface ActionDelegate {
     }
 }
