+++ ide/che-core-ide-app/src/main/java/org/eclipse/che/ide/preferences/pages/extensions/ExtensionManagerViewImpl.ui.xml
@@ -12,7 +12,7 @@
 -->
 <ui:UiBinder xmlns:ui='urn:ui:com.google.gwt.uibinder'
              xmlns:g='urn:import:com.google.gwt.user.client.ui' xmlns:cellview="urn:import:com.google.gwt.user.cellview.client">
-    <ui:style src="org/eclipse/che/ide/api/ui/style.css" type='org.eclipse.che.ide.extension.ExtensionManagerViewImpl.Style'>
+    <ui:style src="org/eclipse/che/ide/api/ui/style.css" type='org.eclipse.che.ide.preferences.pages.extensions.ExtensionManagerViewImpl.Style'>
 
         @eval partBackground org.eclipse.che.ide.api.theme.Style.theme.partBackground();
         @eval separatorColor org.eclipse.che.ide.api.theme.Style.getWindowSeparatorColor();
